Include a generated build artifact for forensics.
Upon triggering a build, Maven will run a script that will fetch
pertinent information from the local environment and bundle it into
a Java class for the pre-compilation phase, which is shortly thereafter
compiled and bundled into the deployment archive.  This will be useful
for cases where we want forensics information from the third-party JAR
once it is deployed in external production systems.